requirements.txt
Django==5.1
djangorestframework==3.15.2
psycopg2-binary==2.9.10
djangorestframework-simplejwt==5.3.1
django-filter==24.3
django-cors-headers==4.6.0
drf-spectacular==0.27.2
Pillow>=10.4.0
python-decouple==3.8
dj-database-url==2.3.0
